refactor: add examples and update clients location
